Replaced the jQuery dependency with Cheerio to solve #28
Node-jQuery causes issue for Windows user because it depends on contexify which requires building some code natively. Cheerio provides the same need functionality for this project whithout the problems ikn Windows and without the overhead of a full DOM implementation.
Furthermore, this resolves an issue with the existing code which has a problem if an image in comments has the same source as an uncommented image, the uncommented image is not modified. The new approach ignoes comments.